feat(MedicalCase): 医案Master-Detail管理与诊断字段重构
## 主要变更
- 新增医案Master-Detail管理视图(MedicalCaseMasterDetailView)
- 重构诊断字段：将现病史/舌诊/脉诊/中医诊断从Consultation移至MedicalCase
- 统一HerbListView控件支持处方药材编辑
- 添加DecocteMethod(煎法)数据库迁移

## OpenSpec提案
- refactor-diagnosis-fields: 诊断字段重构
- unify-herb-list-controls: 统一药材列表控件
- refactor-medicalcase-management: 医案管理重构

## 已知问题
- [BUG] 医案保存时Unit字段验证失败(HTTP 400)
  - 根因：HerbDto.Unit默认值"克"与数据库"g"不一致
  - 状态：调查中，待修复

+++ src/Client/Desktop/Core/LYBT.Desktop.Infrastructure/Interfaces/IPrescriptionSettingsService.cs
@@ -18,9 +18,9 @@ namespace LYBT.Desktop.Infrastructure.Interfaces
         /// <summary>
         /// 计算合并后的剂量
         /// </summary>
-        /// <param name="currentDosage">当前剂量</param>
-        /// <param name="importedDosage">导入剂量</param>
-        /// <returns>根据合并策略计算的剂量</returns>
-        decimal CalculateMergedDosage(decimal currentDosage, decimal importedDosage);
+        /// <param name="currentDosage">当前剂量（整数克）</param>
+        /// <param name="importedDosage">导入剂量（整数克）</param>
+        /// <returns>根据合并策略计算的剂量（整数克）</returns>
+        int CalculateMergedDosage(int currentDosage, int importedDosage);
     }
 }
